requirements.txt
openai
anthropic
google-genai
python-dotenv
httpx
